fix: channel use same socket with game (#28)
